chore(resolvers/schema): modularize schema setup
- use 'merge-graphql-schemas' to modularize schemas
- separate resolvers by entity

[closes #11]
