+++ package.json
@@ -27,6 +27,7 @@
     "dotenv": "^6.2.0",
     "graphql-yoga": "^1.16.7",
     "jsonwebtoken": "^8.4.0",
+    "merge-graphql-schemas": "^1.5.8",
     "pg": "^7.7.1",
     "sequelize": "^4.42.0",
     "sequelize-cli": "^5.4.0"
