+++ src/resolvers/index.js
@@ -1,9 +1,14 @@
-import Query from './Query';
-import Mutation from './Mutation';
+import path from 'path';
+import { fileLoader } from 'merge-graphql-schemas';
 
-const resolvers = {
-  Query,
-  Mutation,
-};
+/* MANUAL APPROACH: Update this file manually with each resolver file */
+// import userResolvers from "./user.resolvers";
+// import welcomeResolvers from "./welcome.resolvers";
+// const resolversArray = [userResolvers, welcomeResolvers];
+
+/*  AUTOMATED APPROACH: Put your resolvers anywhere
+    with ".resolvers.[js/ts]" naming convention */
+const resolvers = fileLoader(path.join(__dirname, './**/*.resolvers.*'));
+console.log('the resolvers', resolvers);
 
 export default resolvers;
