+++ src/resolvers/info.resolvers.js
@@ -1,0 +1,5 @@
+export default {
+  Query: {
+    info: () => 'This is the GraphQL API for Ride My Way',
+  }
+};
