+++ src/schema/info.schema.graphql
@@ -1,0 +1,3 @@
+type Query {
+  info: String!
+}
