--- src/schema/schema.graphql
@@ -1,22 +1,0 @@
-type Query {
-  info: String!
-  getUsers: [User!]
-}
-
-type Mutation {
-  signUp(username: String!, fullName: String!, email: String!, password: String!): AuthPayload!
-  signIn(email: String!, password: String!): AuthPayload!
-}
-
-type User {
-  id: ID!
-  username: String!
-  fullName: String!
-  email: String!
-  phone: String
-}
-
-type AuthPayload {
-  user: User!
-  token: String!
-}
